Revert "controllers: add retries for failed images jobs (#1393)"
This reverts commit 28118640c7d0f70b526d95a1b4da27f20eb93dcf.
